Kevin's branch, Parts 2.3.1, 2.3.5, and respective parts for 2.4 completed (#11)
* Variable names changed for consistency

text_dataset -> dataset in the word_frequency function
frequencies -> frequency_array in the word_frequency function to distinguish from another frequencies variable name

training["posts"] -> X_train
training["emotions"], training["sentiments"] -> y_train
We store the training and testing data as numpy arrays themselves rather than in a dictionary for better clarity in training.

* #4 Base MNB and #5 Top DT

- performance_kevin.txt has the precision, recall, accuracy, and f1 results for Base MNB and Top DT
- Confusion matrices made in seaborn for both Base MNB and Top DT emotion and sentiments classifications

* Added part 2.5 for Base MNB and Top DT.

* Completed 3.1 and 3.2

3.1 and 3.2 are done.
A pie chart of the emotions has also been added. It was written to help with analysis in 4.1.

* 3.3 Completed

Now posts can have embeddings which are the average of the word embeddings of the words in the post.

* Completed 3.4

hit rates for the training and testing sets

* Fixed data models, added confusion matrices, updated performance

data models were showing incorrect results because the emotions classifiers and sentiments classifiers were calling .fit() next to each other, therefore, the data model would then be set to fit for the last one called. In doing so, test emotions were being predicted with a sentiments fitted model. All confusion matrices and performance results in the .txt have been updated as a result. The notebook has also been reformatted to account for this issue.

* added base mnb sentiments classication report in perfomance.txt

forgot this in first commit

* fixed mislabeling on confusion matrices

the labels needed to be sorted in alphabetical order, they should all be fine now.

* renamed notebook

* deleted the graphs

it has no use in my part
trimming my notebook to make it more clean.

Co-authored-by: Vatanak So <76747982+vatanium-lite@users.noreply.github.com>